- Fix for Multi-selection in day Page.
Handler for mouse clicking and PerformSelection method were moved to EntryViewBase class. After debuggin this handler, it seems to be called multiple times while clicking anywhere inside the view and of course it happens due to each control that shapes the entire view has this handler, causing that clicked entry get added and removed from the Selection list over and over again.

This situation was preventing that Multi-selection works correctly in the day page, but having that logic in the EntryViewBase class, it give us the expected performance.