Update dependency versions and module references
Updated the versions of ical4j, slf4j, and commons-codec dependencies in the pom.xml file. Also modified the requires statements in module-info.java files to reflect the new ical4j module name. These changes ensure that the project is using the latest stable versions of the dependencies and that module dependencies are correctly specified.